Подготавливает отдельные шаблоны-блоки постов и показывает информацию о посте на отдельной странице
