+++ config.php
@@ -5,3 +5,5 @@
         'password' => '',
         'database' => 'readme',
     ];
+
+    date_default_timezone_set("Europe/Moscow");
