+++ data.php
@@ -5,6 +5,3 @@
     define('WEEK', 7 * DAY);
     define('MONTH', 5 * WEEK);
     define('DATE_TITLE', 'd.m.Y H:i');
-
-    $is_auth = rand(0, 1);
-    $user_name = 'Игорь Влащенко';
