+++ templates/post-photo.php
@@ -1,0 +1,3 @@
+<div class="post-details__image-wrapper post-photo__image-wrapper">
+	<img src="img/<?= $post['img_url']; ?>" alt="Фото от пользователя" width="760" height="507">
+</div>
