+++ templates/post-preview.php
@@ -1,92 +1,146 @@
-<article class="popular__post post post-<?= $value['class']; ?>">
-    <header class="post__header">
-        <h2><a href="post.php?id=<?= $value['post_id'] ?>"><?= htmlspecialchars($value['title']); ?></a></h2>
-    </header>
-    <div class="post__main">
-        <?php if($value['class'] == 'quote'): ?>
-            <blockquote>
-                <p>
-                    <?= htmlspecialchars($value['text_content']);?>
-                </p>
-                <cite>Неизвестный Автор</cite>
-            </blockquote>
-        <?php elseif ($value['class'] == 'text'): ?>
-            <?php $arr_content = cut_text(htmlspecialchars($value['text_content'])) ?>
-            <?php if ($arr_content[1]): ?>
-                <p>
-                    <?=$arr_content[0] . '...'; ?>
-                </p>
-                <a class="post-text__more-link" href="#">Читать далее</a>
-                <?php else: ?>
-                <p>
-                    <?= htmlspecialchars($value['text_content']); ?>
-                </p>
-            <?php endif; ?>
-        <?php elseif ($value['class'] == 'photo'): ?>
-            <div class="post-photo__image-wrapper">
-                <img src="img/<?= $value['img_url']; ?>" alt="Фото от пользователя" width="360" height="240">
-            </div>
-        <?php elseif ($value['class'] == 'video'): ?>
-            <div class="post-video__block">
-                <div class="post-video__preview">
-                    <?= embed_youtube_cover($value['video_url']); ?>
-                </div>
-                <a href="post-details.html" class="post-video__play-big button">
-                    <svg class="post-video__play-big-icon" width="14" height="14">
-                        <use xlink:href="#icon-video-play-big"></use>
-                    </svg>
-                    <span class="visually-hidden">Запустить проигрыватель</span>
-                </a>
-            </div>
-        <?php elseif ($value['class'] == 'link'): ?>
-            <div class="post-link__wrapper">
-                <a class="post-link__external" href="http://<?= $value['site_url']; ?>" title="Перейти по ссылке">
-                    <div class="post-link__info-wrapper">
-                        <div class="post-link__icon-wrapper">
-                            <img src="https://www.google.com/s2/favicons?domain=vitadental.ru" alt="Иконка">
-                        </div>
-                        <div class="post-link__info">
-                            <h3><?= htmlspecialchars($value['title']); ?></h3>
-                        </div>
+<main class="page__main page__main--publication">
+    <div class="container">
+    <h1 class="page__title page__title--publication"><?= $post['title']; ?></h1>
+    <section class="post-details">
+        <h2 class="visually-hidden">Публикация</h2>
+        <div class="post-details__wrapper post-text">
+
+            <div class="post-details__main-block post post--details">
+                <?php
+                    print(include_template('post-' . $post['class'] . '.php', [
+                        'post' => $post,
+                    ]));
+                ?>
+                <div class="post__indicators">
+                    <div class="post__buttons">
+                        <a class="post__indicator post__indicator--likes button" href="#" title="Лайк">
+                        <svg class="post__indicator-icon" width="20" height="17">
+                            <use xlink:href="#icon-heart"></use>
+                        </svg>
+                        <svg class="post__indicator-icon post__indicator-icon--like-active" width="20" height="17">
+                            <use xlink:href="#icon-heart-active"></use>
+                        </svg>
+                        <span><?= $post['likes']; ?></span>
+                        <span class="visually-hidden">количество лайков</span>
+                        </a>
+                        <a class="post__indicator post__indicator--comments button" href="#" title="Комментарии">
+                        <svg class="post__indicator-icon" width="19" height="17">
+                            <use xlink:href="#icon-comment"></use>
+                        </svg>
+                        <span>25</span>
+                        <span class="visually-hidden">количество комментариев</span>
+                        </a>
+                        <a class="post__indicator post__indicator--repost button" href="#" title="Репост">
+                        <svg class="post__indicator-icon" width="19" height="17">
+                            <use xlink:href="#icon-repost"></use>
+                        </svg>
+                        <span>5</span>
+                        <span class="visually-hidden">количество репостов</span>
+                        </a>
                     </div>
-                    <span><?= htmlspecialchars($value['site_url']); ?></span>
-                </a>
+                    <span class="post__view"><?= $post['view_counter']; ?> <?= get_noun_plural_form($post['view_counter'], 'просмотр', 'просмотра', 'просмотров'); ?></span>
+                </div>
+                <ul class="post__tags">
+                <li><a href="#">#nature</a></li>
+                <li><a href="#">#globe</a></li>
+                <li><a href="#">#photooftheday</a></li>
+                <li><a href="#">#canon</a></li>
+                <li><a href="#">#landscape</a></li>
+                <li><a href="#">#щикарныйвид</a></li>
+                </ul>
+                <div class="comments">
+                <form class="comments__form form" action="#" method="post">
+                    <div class="comments__my-avatar">
+                    <img class="comments__picture" src="img/userpic-medium.jpg" alt="Аватар пользователя">
+                    </div>
+                    <div class="form__input-section form__input-section--error">
+                    <textarea class="comments__textarea form__textarea form__input" placeholder="Ваш комментарий"></textarea>
+                    <label class="visually-hidden">Ваш комментарий</label>
+                    <button class="form__error-button button" type="button">!</button>
+                    <div class="form__error-text">
+                        <h3 class="form__error-title">Ошибка валидации</h3>
+                        <p class="form__error-desc">Это поле обязательно к заполнению</p>
+                    </div>
+                    </div>
+                    <button class="comments__submit button button--green" type="submit">Отправить</button>
+                </form>
+                <div class="comments__list-wrapper">
+                    <ul class="comments__list">
+                    <li class="comments__item user">
+                        <div class="comments__avatar">
+                        <a class="user__avatar-link" href="#">
+                            <img class="comments__picture" src="img/userpic-larisa.jpg" alt="Аватар пользователя">
+                        </a>
+                        </div>
+                        <div class="comments__info">
+                        <div class="comments__name-wrapper">
+                            <a class="comments__user-name" href="#">
+                            <span>Лариса Роговая</span>
+                            </a>
+                            <time class="comments__time" datetime="2019-03-20">1 ч назад</time>
+                        </div>
+                        <p class="comments__text">
+                            Красота!!!1!
+                        </p>
+                        </div>
+                    </li>
+                    <li class="comments__item user">
+                        <div class="comments__avatar">
+                        <a class="user__avatar-link" href="#">
+                            <img class="comments__picture" src="img/userpic-larisa.jpg" alt="Аватар пользователя">
+                        </a>
+                        </div>
+                        <div class="comments__info">
+                        <div class="comments__name-wrapper">
+                            <a class="comments__user-name" href="#">
+                            <span>Лариса Роговая</span>
+                            </a>
+                            <time class="comments__time" datetime="2019-03-18">2 дня назад</time>
+                        </div>
+                        <p class="comments__text">
+                            Озеро Байкал – огромное древнее озеро в горах Сибири к северу от монгольской границы. Байкал считается самым глубоким озером в мире. Он окружен сетью пешеходных маршрутов, называемых Большой байкальской тропой. Деревня Листвянка, расположенная на западном берегу озера, – популярная отправная точка для летних экскурсий. Зимой здесь можно кататься на коньках и собачьих упряжках.
+                        </p>
+                        </div>
+                    </li>
+                    </ul>
+                    <a class="comments__more-link" href="#">
+                    <span>Показать все комментарии</span>
+                    <sup class="comments__amount">45</sup>
+                    </a>
+                </div>
+                </div>
             </div>
-        <?php endif; ?>
+            <!--  -->
+            <div class="post-details__user user">
+                <div class="post-details__user-info user__info">
+                <div class="post-details__avatar user__avatar">
+                    <a class="post-details__avatar-link user__avatar-link" href="#">
+                        <img class="post-details__picture user__picture" src="img/<?= $post['avatar_url'] ?>" alt="Аватар пользователя">
+                    </a>
+                </div>
+                <div class="post-details__name-wrapper user__name-wrapper">
+                    <a class="post-details__name user__name" href="#">
+                    <span><?= $post['author']; ?></span>
+                    </a>
+                    <time class="post-details__time user__time" datetime="2014-03-20">5 лет на сайте</time>
+                </div>
+                </div>
+                <div class="post-details__rating user__rating">
+                <p class="post-details__rating-item user__rating-item user__rating-item--subscribers">
+                    <span class="post-details__rating-amount user__rating-amount"><?= $subscribers['count_sub']; ?></span>
+                    <span class="post-details__rating-text user__rating-text"><?= get_noun_plural_form($subscribers['count_sub'], 'подписчик', 'подписчика', 'подписчиков'); ?></span>
+                </p>
+                <p class="post-details__rating-item user__rating-item user__rating-item--publications">
+                    <span class="post-details__rating-amount user__rating-amount"><?= $posting['count_posts']; ?></span>
+                    <span class="post-details__rating-text user__rating-text"><?= get_noun_plural_form($posting['count_posts'], 'публикация', 'публикации', 'публикаций'); ?></span>
+                </p>
+                </div>
+                <div class="post-details__user-buttons user__buttons">
+                <button class="user__button user__button--subscription button button--main" type="button">Подписаться</button>
+                <a class="user__button user__button--writing button button--green" href="#">Сообщение</a>
+                </div>
+            </div>
+        </div>
+    </section>
     </div>
-    <footer class="post__footer">
-        <div class="post__author">
-            <a class="post__author-link" href="#" title="Автор">
-                <div class="post__avatar-wrapper">
-                    <img class="post__author-avatar" src="img/<?= $value['avatar_url']; ?>" alt="Аватар пользователя">
-                </div>
-                <div class="post__info">
-                    <b class="post__author-name"><?= htmlspecialchars($value['author']); ?></b>
-                    <?php $public_date = generate_random_date($key); ?>
-                    <time class="post__time" datetime="<?= $public_date ?>" title="<?= format_date($public_date, DATE_TITLE); ?>"><?= get_diff_time_public_post($public_date); ?></time>
-                </div>
-            </a>
-        </div>
-        <div class="post__indicators">
-            <div class="post__buttons">
-                <a class="post__indicator post__indicator--likes button" href="#" title="Лайк">
-                    <svg class="post__indicator-icon" width="20" height="17">
-                        <use xlink:href="#icon-heart"></use>
-                    </svg>
-                    <svg class="post__indicator-icon post__indicator-icon--like-active" width="20" height="17">
-                        <use xlink:href="#icon-heart-active"></use>
-                    </svg>
-                    <span><?= $value['likes']; ?></span>
-                    <span class="visually-hidden">количество лайков</span>
-                </a>
-                <a class="post__indicator post__indicator--comments button" href="#" title="Комментарии">
-                    <svg class="post__indicator-icon" width="19" height="17">
-                        <use xlink:href="#icon-comment"></use>
-                    </svg>
-                    <span>0</span>
-                    <span class="visually-hidden">количество комментариев</span>
-                </a>
-            </div>
-        </div>
-    </footer>
-</article>
+</main>
