+++ templates/post-quote.php
@@ -1,0 +1,10 @@
+<div class="post-details__image-wrapper post-quote">
+<div class="post__main">
+  <blockquote>
+    <p>
+    <?= $post['text_content']; ?>
+    </p>
+    <cite><?= $post['author_quote']; ?></cite>
+  </blockquote>
+</div>
+</div>
