+++ templates/post-text.php
@@ -1,0 +1,7 @@
+<div class="post-details__image-wrapper post-text">
+  <div class="post__main">
+    <p>
+      <?= $post['text_content']; ?>
+    </p>
+  </div>
+</div>
