+++ templates/post-video.php
@@ -1,0 +1,3 @@
+<div class="post-details__image-wrapper post-photo__image-wrapper">
+  <?= embed_youtube_video($post['video_url']); ?>
+</div>
